pyproject.toml
[project]
name = "sukuna-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elevenlabs>=2.16.0",
    "groq>=0.32.0",
    "gtts>=2.5.4",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.8",
    "opencv-python>=4.12.0.88",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "speechrecognition>=3.14.3",
]
